requirements.txt
    asgiref==3.2.7
Django==3.0.4
djangorestframework==3.11.0
django-cors-headers==3.2.1
djangorestframework-jwt==1.11.0
Jinja2==2.11.1
MarkupSafe==1.1.1
mysqlclient==1.4.6
pytz==2019.3
sqlparse==0.3.1
django-bootstrap4==1.1.1